Remove underscores in EPS/PS/PDF/SVGwriter options
Remove the underscore in the arguments strip_fonts, text_as_path,
mesh_as_bitmap, mesh_as_bitmap_resolution (new: stripfonts, textaspath,
meshasbitmap and meshasbitmapresolution) of EPSwriter, PSwriter,
PDFwriter and SVGwriter. This change has been made in order to remove a
potential ambiquity when using the prefix write_ to set these options,
in particular since it is planned to enable such prefixes in more parts
of PyX in the future.

The old argument names are still supported but a deprecation warning is
issued.