fix t1 extension with Python 3.10
For all # variants of formats (s#, y#, etc.), the macro PY_SSIZE_T_CLEAN
must be defined before including Python.h. On Python 3.9 and older, the
type of the length argument is Py_ssize_t if the PY_SSIZE_T_CLEAN macro
is defined, or int otherwise. [https://docs.python.org/3.10/c-api/arg.html]

So, on Python 3.9, PyX worked because PY_SSIZE_T_CLEAN was not defined
(even though it was recommended), so that the deprecated return type of
int (for the string length) was used.

Python 3.10 finally pulls the trigger on the ints, so finally adjust to
that announced change.